Rename e2e_test.py to workflow_test.py
