+++ stoobly_agent/test/app/cli/scaffold/docker/workflow_test.py
